# openapi.yaml
openapi: 3.0.3
info:
  title: BLM Spider API
  version: 1.0.0
  description: |
    An API for exploring Bureau of Land Management (BLM) dispersed campsites across the U.S. This project combines a clustered Node.js server with a high-performance web crawler to extract and serve structured campsite data from the BLM website. Data is enriched with offline geospatial boundaries and made available through RESTful endpoints with support for filtering and pagination. The crawler is optimized for speed and reliability using parallel page fetching, persistent connections, and retry logic. Updates run automatically each week via a GitHub Actions cron job to keep data fresh.
    > *Disclaimer: Not all results are guaranteed to be valid campsites. Best-effort filters are applied, but users should verify site status with the BLM before visiting.*

    ### CORS Policy 
    All endpoints support [CORS](https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS) (Cross-Origin Resource Sharing). This means you can call the API directly from web browsers and client-side applications without additional configuration.

    ### Rate Limiting
    To ensure fair use and prevent abuse, the API enforces rate limiting:
    - Maximum **100 requests per 15 minutes** per IP address.
    - If you exceed this limit, you will receive a `429 Too Many Requests` error.  
    - Rate limits are enforced globally across all endpoints.
  contact:
    name: Bailey Dunning
    url: https://github.com/baileydunning/blm-spider
servers:
- url: https://blm-spider.onrender.com/
  description: Production server
- url: http://localhost:3000
  description: Local development server

tags:
- name: Campsites
  description: Operations for retrieving BLM campsite data

paths:
  /api/v1/campsites:
    get:
      tags: [ Campsites ]
      summary: Retrieve all campsites
      description: |
        Returns a list of BLM campsites. Supports optional filtering by state and pagination.

        **Usage notes:**
        - Filter by state using the `state` query parameter (case-insensitive, e.g. `state=Wyoming`)
        - Use `limit` and `offset` for pagination
      parameters:
      - in: query
        name: state
        schema:
          type: string
        description: Filter campsites by U.S. state (full name, e.g. "Wyoming")
      - in: query
        name: activities
        schema:
          type: string
        description: Comma-separated list of activities to filter by (e.g. "camping, hiking")
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
        description: Maximum number of results to return
      - in: query
        name: offset
        schema:
          type: integer
          minimum: 0
        description: Number of results to skip before starting to return results
      responses:
        '200':
          description: A list of campsites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campsite'
              examples:
                example:
                  summary: Example response
                  value:
                  - id: "28430a4f-ce40-478d-a7b1-0c23d9c49dbb"
                    name: "Middle Fork of the Powder River Campground"
                    url: "https://www.blm.gov/visit/middle-fork-powder-river-campground-0"
                    description: "Astride a blue-ribbon trout stream, this remote and picturesque campground features five camping sites with fire rings, restroom and drinking water. Use of the area is free with a 14-day limit on camping (as on all public lands). The Middle Fork of the Powder River Campground, located at an elevation of 7,500 feet, is open from May through September."
                    lat: 43.579304
                    lng: -107.140476
                    state: "Wyoming"
                    mapLink: "https://www.openstreetmap.org/export/embed.html?bbox=-107.15047600000001,43.569304,-107.130476,43.589304&layer=mapnik&marker=43.579304,-107.140476"
                    directions: "From Ten Sleep, the area is reached by driving 20 miles south on State Highway 434 to Big Trails. From there turn left on the graveled Dry Farm Road and drive about 13 miles to the Hazelton Road. Turn right and drive about 17 miles to the campground. For other routes to the campground, check with the BLM offices in Worland, Buffalo, or Casper."
                    activities: [ "CAMPING" ]
                    source: "BLM"
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  value:
                    error: "Invalid 'limit' parameter"
        '429':
          description: Too Many Requests – rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                rate_limited:
                  value:
                    error: "Too many requests. Please try again later."
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  value:
                    error: "Internal server error"

  /api/v1/campsites/{id}:
    get:
      tags: [ Campsites ]
      summary: Get a single campsite by ID
      description: Retrieves detailed information for a specific campsite by its unique ID.
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: The unique ID of the campsite (UUID)
      responses:
        '200':
          description: Campsite found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campsite'
              examples:
                example:
                  summary: Example response
                  value:
                    id: "5a1e40b5-4aec-4624-94e0-6f9e757a61de"
                    name: "Valley of the Giants Outstanding Natural Area"
                    url: "https://www.blm.gov/visit/valley-giants-outstanding-natural-area"
                    description: "Some of the largest Douglas firs and Western Hemlocks in the Oregon Coast Range can be found in the Bureau of Land Management’s Valley of the Giants. Many of these trees are twenty feet in circumference, 200 feet tall, and 400 to 450 years old."
                    lat: 44.936289
                    lng: -123.715592
                    state: "Oregon"
                    mapLink: "https://www.openstreetmap.org/export/embed.html?bbox=-123.725592,44.926289000000004,-123.705592,44.946289&layer=mapnik&marker=44.936289,-123.715592"
                    directions: "Visitors should contact the Northwest Oregon District Office for directions and current conditions. Route closed when public fire restrictions are in effect."
                    fees: "No fees."
                    images:
                    - src: "https://cdn.recreation.gov/public/2020/06/09/21/21/10057631_a27238fc-a182-4009-84dc-9955a7e1433e.jpeg"
                      alt: "Tall trees on display in the Valley of the Giants."
                      credit: "Bureau of Land Management"
                    source: "BLM"
        '404':
          description: Campsite not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  value:
                    error: "Campsite not found"
        '429':
          description: Too Many Requests – rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                rate_limited:
                  value:
                    error: "Too many requests. Please try again later."
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  value:
                    error: "Internal server error"

components:
  schemas:
    Campsite:
      type: object
      required:
      - id
      - name
      - url
      - lat
      - lng
      - state
      - mapLink
      - source
      properties:
        id:
          type: string
          description: Unique campsite identifier (UUID v4)
        name:
          type: string
          description: Official name of the campsite or area
        url:
          type: string
          description: Source BLM webpage URL for the campsite
        lat:
          type: number
          format: float
          description: Latitude in decimal degrees (WGS84)
        lng:
          type: number
          format: float
          description: Longitude in decimal degrees (WGS84)
        state:
          type: string
          description: U.S. state name (inferred from coordinates if not present on the page)
        mapLink:
          type: string
          description: URL for an embeddable OpenStreetMap or ArcGIS map of the campsite location
        description:
          type: string
          nullable: true
          description: Overview or description of the campsite, including amenities and setting
        directions:
          type: string
          nullable: true
          description: Driving or access directions to the campsite
        activities:
          type: array
          items:
            type: string
          nullable: true
          description: List of activities available at or near the campsite (e.g., CAMPING, HIKING, FISHING)
        campgrounds:
          type: array
          items:
            type: string
          nullable: true
          description: Associated campground names (if the site is part of a larger area)
        wildlife:
          type: array
          items:
            type: string
          nullable: true
          description: Common wildlife species in the area
        fees:
          type: string
          nullable: true
          description: Fee information, if any (e.g., "No fees", "$10/night")
        stayLimit:
          type: string
          nullable: true
          description: Maximum duration allowed for camping (e.g., "14 days")
        images:
          type: array
          items:
            $ref: '#/components/schemas/CampsiteImage'
          nullable: true
          description: Array of photos of the campsite, with alt text and credit
        source:
          type: string
          enum: [ BLM ]
          description: Data source for the campsite (always "BLM")

    CampsiteImage:
      type: object
      required:
      - src
      properties:
        src:
          type: string
          description: Image URL
        alt:
          type: string
          nullable: true
          description: Alternative text
        credit:
          type: string
          nullable: true
          description: Image credit

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message